requirements.txt
pillow==10.2.0
django-filter==24.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1